+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Decoded api key byte string [#2482](https://github.com/opendatateam/udata/pull/2482)
 - Removed now useless metric fetching [#2482](https://github.com/opendatateam/udata/pull/2484)
+- Fix bug in harvester's cron schedule [#2493](https://github.com/opendatateam/udata/pull/2493)
 
 ## 2.1.0 (2020-05-12)